Inventory Management System(Ass 2)/requirements.txt
# Core Django
Django==5.0.4
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1

# Database
psycopg2-binary==2.9.9

# CORS and Security
django-cors-headers==4.3.1

# Authentication and Authorization
django-otp==1.5.4
PyJWT==2.8.0
django-rest-passwordreset==1.4.1

# API Documentation
drf-yasg==1.21.7
pyyaml==6.0.1
uritemplate==4.1.1

# AWS Services
boto3==1.34.98
botocore==1.34.98
s3transfer==0.10.1

# Celery (Task Queue)
celery==5.4.0
flower==2.0.1
redis==5.0.4
django-redis==5.4.0

# Server
gunicorn==22.0.0
whitenoise==6.6.0

# Utilities
django-extensions==3.2.3
django-filter==24.2
python-dateutil==2.9.0.post0
pytz==2024.1
timeago==1.0.16
wget==3.2
qrcode==7.4.2

# Image Processing
pillow==10.3.0
pillow-heif==0.16.0

# Firebase
firebase-admin==6.5.0

# Google Maps
googlemaps==4.10.0

# MongoDB
pymongo==4.7.1

# Data Import/Export
django-import-export==4.0.0

# Development Tools (should be in dev-requirements.txt)
django-debug-toolbar==4.3.0
flake8==7.0.0
mccabe==0.7.0
pycodestyle==2.11.1
pyflakes==3.2.0

# Core Python Dependencies
asgiref==3.8.1
jmespath==1.0.1
six==1.16.0
sqlparse==0.5.0
urllib3==2.2.1
